requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
colorama==0.4.6
distro==1.9.0
executing==2.1.0
Farama-Notifications==0.0.4
gymnasium==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
icecream==2.1.3
idna==3.10
jiter==0.6.1
numpy==2.1.2
openai==1.52.1
orjson==3.10.9
poke-env==0.8.1
pydantic==2.9.2
pydantic-core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
websockets==13.1
pandas==2.2.3